PointCloud.py
import numpy as np
import os
import trimesh
import random
import torch 
from torchvision import transforms,utils
from torch.utils.data import Dataset, DataLoader

def read_stl(path_stl):
  mesh = trimesh.load(path_stl)
  return mesh.vertices,mesh.faces

class PointSampler(object):
    def __init__(self, output_size):
        assert isinstance(output_size, int)
        self.output_size = output_size
    
    def triangle_area(self, pt1, pt2, pt3):
        side_a = np.linalg.norm(pt1 - pt2)
        side_b = np.linalg.norm(pt2 - pt3)
        side_c = np.linalg.norm(pt3 - pt1)
        s = 0.5 * ( side_a + side_b + side_c)
        return max(s * (s - side_a) * (s - side_b) * (s - side_c), 0)**0.5

    def sample_point(self, pt1, pt2, pt3):
        s, t = sorted([random.random(), random.random()])
        f = lambda i: s * pt1[i] + (t-s)*pt2[i] + (1-t)*pt3[i]
        return (f(0), f(1), f(2))
        
    
    def __call__(self, mesh):
        verts, faces = mesh
        verts = np.array(verts)
        areas = np.zeros((len(faces)))

        for i in range(len(areas)):
            areas[i] = (self.triangle_area(verts[faces[i][0]],
                                           verts[faces[i][1]],
                                           verts[faces[i][2]]))
            
        sampled_faces = (random.choices(faces, 
                                      weights=areas,
                                      cum_weights=None,
                                      k=self.output_size))
        
        sampled_points = np.zeros((self.output_size, 3))

        for i in range(len(sampled_faces)):
            sampled_points[i] = (self.sample_point(verts[sampled_faces[i][0]],
                                                   verts[sampled_faces[i][1]],
                                                   verts[sampled_faces[i][2]]))
        
        return sampled_points

class ToTensor(object):
    def __call__(self, pointcloud):
        assert len(pointcloud.shape)==2

        return torch.from_numpy(pointcloud)

def default_transforms():
    return transforms.Compose([
                                PointSampler(8000),
                                ToTensor()
                              ])
class PointCloudData(Dataset):
    def __init__(self, file_name, valid=False, folder="train", transform=default_transforms()):
        self.files = []
        self.transforms = transform if not valid else default_transforms()
        if file_name.endswith('.stl'):
                    sample = {}
                    sample['pcd_path'] = file_name
                    # sample['category'] = category
                    self.files.append(sample)

    def __len__(self):
        return len(self.files)

    def __preproc__(self, file):
        # print(file.name)
        verts, faces = read_stl(file.name)
        if self.transforms:
            pointcloud = self.transforms((verts, faces))
        return pointcloud

    def __getitem__(self, idx):
        pcd_path = self.files[idx]['pcd_path']
        # category = self.files[idx]['category']
        with open(pcd_path, 'r') as f:
            pointcloud = self.__preproc__(f)
        return {'pointcloud': pointcloud}

train_transforms = transforms.Compose([
                    PointSampler(4096),
                    ToTensor()
                    ])
def  spine_ds(path):                    
    full_spine_ds = PointCloudData(path, transform=train_transforms)
    full_spine_loader = DataLoader(dataset=full_spine_ds)
    return(full_spine_loader)