requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
backcall=0.2.0=py_0
blas=1.0=mkl
bzip2=1.0.8=he774522_0
ca-certificates=2020.10.14=0
certifi=2020.11.8=py38haa95532_0
colorama=0.4.3=py_0
cpuonly=1.0=0
curl=7.71.1=h2a8f88b_1
decorator=4.4.2=py_0
double-conversion=3.1.5=ha925a31_1
eigen=3.3.7=h74a9793_0
expat=2.2.9=h33f27b4_2
ffmpeg=4.3.1=ha925a31_0
freetype=2.10.2=hd328e21_0
gl2ps=1.4.2=h2c487a6_0
glew=2.1.0=he025d50_0
hdf4=4.2.13=h712560f_2
hdf5=1.10.6=nompi_he0bbb20_101
intel-openmp=2020.1=216
ipykernel=5.3.4=py38h5ca1d4c_0
ipython=7.16.1=py38h5ca1d4c_0
ipython_genutils=0.2.0=py38_0
itk=5.1.1=py38h32f6830_2
itk-core=5.1.1=pypi_0
itk-filtering=5.1.1=pypi_0
itk-numerics=5.1.1=pypi_0
itk-registration=5.1.1=pypi_0
jedi=0.15.2=py38_0
jpeg=9d=he774522_0
jsoncpp=1.8.4=h74a9793_0
jupyter_client=6.1.6=py_0
jupyter_core=4.6.3=py38_0
krb5=1.18.2=hc04afaa_0
libcurl=7.71.1=h2a8f88b_1
libiconv=1.15=h1df5818_7
libnetcdf=4.7.4=nompi_h256d12c_105
libogg=1.3.2=he774522_0
libpng=1.6.37=h2a8f88b_0
libsodium=1.0.18=h62dcd97_0
libssh2=1.9.0=h7a1dbc1_1
libtheora=1.1.1=h62dcd97_1003
libtiff=4.1.0=h56a325e_1
libuv=1.40.0=he774522_0
libxml2=2.9.10=h464c3ec_1
loguru=0.5.0=py38h32f6830_0
lz4-c=1.9.2=h62dcd97_1
medpy=0.4.0=pypi_0
mkl=2020.1=216
mkl-service=2.3.0=py38hb782905_0
mkl_fft=1.1.0=py38h45dec08_0
mkl_random=1.1.1=py38h47e9c7a_0
ninja=1.10.1=py38h7ef1ec2_0
nose=1.3.7=pypi_0
numpy=1.19.1=py38h5510c5b_0
numpy-base=1.19.1=py38ha3acd2a_0
numpy-stl=2.11.2=py38h4b8e87e_0
olefile=0.46=py_0
openssl=1.1.1h=he774522_0
parso=0.5.2=py_0
pickleshare=0.7.5=py38_1000
pillow=8.0.1=py38h7011068_0
pip=20.2.2=py38_0
plotly=4.9.0=py_0
proj=7.0.0=haa36216_5
prompt-toolkit=3.0.5=py_0
pugixml=1.10=h6538335_1
pygments=2.6.1=py_0
pymesh=1.0.2=pypi_0
python=3.8.5=he1778fa_0
python-dateutil=2.8.1=py_0
python-utils=2.3.0=py38_0
python_abi=3.8=1_cp38
pytorch=1.7.0=py3.8_cpu_0
pywin32=227=py38he774522_1
pyzmq=19.0.1=py38ha925a31_1
retrying=1.3.3=py_2
scipy=1.5.2=pypi_0
setuptools=49.6.0=py38_0
simpleitk=2.0.0=py38ha925a31_0
six=1.15.0=py_0
tbb=2020.0=h74a9793_0
tbb-devel=2020.0=h74a9793_0
tk=8.6.10=he774522_0
torchaudio=0.7.0=py38
torchvision=0.8.1=py38_cpu
tornado=6.0.4=py38he774522_1
traitlets=4.3.3=py38_0
trimesh=3.8.1=pyh9f0ad1d_0
typing_extensions=3.7.4.3=py_0
utfcpp=3.1.1=0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_3
vtk=9.0.1=no_osmesa_py38hf2580a8_100
wcwidth=0.2.5=py_0
wheel=0.34.2=py38_0
win32_setctime=1.0.1=py_0
wincertstore=0.2=py38_0
xz=5.2.5=h62dcd97_0
zeromq=4.3.2=ha925a31_2
zlib=1.2.11=h62dcd97_4
zstd=1.4.5=h04227a9_0
